backend/requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
SQLAlchemy
alembic
# For GPU support, install torch and torchvision with CUDA 12.1 wheels from the official PyTorch index
# To install with CUDA support, use:
# pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu121

torch
torchvision
pymupdf
msal
easyocr
pytesseract
pdf2image
pillow
opencv-python-headless
cachetools
pydantic[email]
aiofiles